Moved peer parsing code back into cache_cf.cc
