+++ src/cache_cf.cc
@@ -24,6 +24,7 @@
 #include "base/PackableStream.h"
 #include "base/RunnersRegistry.h"
 #include "cache_cf.h"
+#include "CachePeer.h"
 #include "CachePeers.h"
 #include "ConfigOption.h"
 #include "ConfigParser.h"
@@ -138,6 +139,8 @@ static void parse_ecap_service_type(Adaptation::Ecap::Config *);
 static void dump_ecap_service_type(StoreEntry *, const char *, const Adaptation::Ecap::Config &);
 static void free_ecap_service_type(Adaptation::Ecap::Config *);
 #endif
+
+static peer_t parseNeighborType(const char *s);
 
 static const char *const T_NANOSECOND_STR = "nanosecond";
 static const char *const T_MICROSECOND_STR = "microsecond";
@@ -2038,11 +2041,372 @@ parse_cachedir(Store::DiskConfig *swap)
     Store::Disks::Parse(*swap);
 }
 
+static const char *
+peer_type_str(const peer_t type)
+{
+    const char * result;
+
+    switch (type) {
+
+    case PEER_PARENT:
+        result = "parent";
+        break;
+
+    case PEER_SIBLING:
+        result = "sibling";
+        break;
+
+    case PEER_MULTICAST:
+        result = "multicast";
+        break;
+
+    default:
+        result = "unknown";
+        break;
+    }
+
+    return result;
+}
+
 static void
 dump_peer(StoreEntry * entry, const char *name, const CachePeers *peers)
 {
-    if (peers)
-        peers->dump(entry, name);
+    if (!peers)
+        return;
+
+    NeighborTypeDomainList *t;
+    LOCAL_ARRAY(char, xname, 128);
+
+    for (const auto &peer: (*peers)) {
+        const auto p = peer.get();
+        storeAppendPrintf(entry, "%s %s %s %d %d name=%s",
+                          name,
+                          p->host,
+                          neighborTypeStr(p),
+                          p->http_port,
+                          p->icp.port,
+                          p->name);
+        dump_peer_options(entry, p);
+
+        if (p->access) {
+            snprintf(xname, 128, "cache_peer_access %s", p->name);
+            dump_acl_access(entry, xname, p->access);
+        }
+
+        for (t = p->typelist; t; t = t->next) {
+            storeAppendPrintf(entry, "neighbor_type_domain %s %s %s\n",
+                              p->host,
+                              peer_type_str(t->type),
+                              t->domain);
+        }
+    }
+}
+
+/**
+ * utility function to prevent getservbyname() being called with a numeric value
+ * on Windows at least it returns garage results.
+ */
+static bool
+isUnsignedNumeric(const char *str, size_t len)
+{
+    if (len < 1) return false;
+
+    for (; len >0 && *str; ++str, --len) {
+        if (! isdigit(*str))
+            return false;
+    }
+    return true;
+}
+
+/**
+ \param proto   'tcp' or 'udp' for protocol
+ \returns       Port the named service is supposed to be listening on.
+ */
+static unsigned short
+GetService(const char *proto)
+{
+    struct servent *port = nullptr;
+    /** Parses a port number or service name from the squid.conf */
+    char *token = ConfigParser::NextToken();
+    if (token == nullptr)
+        throw TextException("cache_peer port is missing", Here());
+
+    /** Returns either the service port number from /etc/services */
+    if ( !isUnsignedNumeric(token, strlen(token)) )
+        port = getservbyname(token, proto);
+    if (port != nullptr) {
+        return ntohs((unsigned short)port->s_port);
+    }
+    /** Or a numeric translation of the config text. */
+    return xatos(token);
+}
+
+/**
+ \returns       Port the named TCP service is supposed to be listening on.
+ \copydoc GetService(const char *proto)
+ */
+inline unsigned short
+GetTcpService(void)
+{
+    return GetService("tcp");
+}
+
+/**
+ \returns       Port the named UDP service is supposed to be listening on.
+ \copydoc GetService(const char *proto)
+ */
+inline unsigned short
+GetUdpService(void)
+{
+    return GetService("udp");
+}
+
+static void
+parse_peer(CachePeers **peers)
+{
+    char *host_str = ConfigParser::NextToken();
+    if (!host_str) {
+        self_destruct();
+        return;
+    }
+
+    char *token = ConfigParser::NextToken();
+    if (!token) {
+        self_destruct();
+        return;
+    }
+
+    const auto p = new CachePeer(host_str);
+
+    p->type = parseNeighborType(token);
+
+    if (p->type == PEER_MULTICAST) {
+        p->options.no_digest = true;
+        p->options.no_netdb_exchange = true;
+    }
+
+    p->http_port = GetTcpService();
+
+    if (!p->http_port) {
+        delete p;
+        self_destruct();
+        return;
+    }
+
+    p->icp.port = GetUdpService();
+
+    while ((token = ConfigParser::NextToken())) {
+        if (!strcmp(token, "proxy-only")) {
+            p->options.proxy_only = true;
+        } else if (!strcmp(token, "no-query")) {
+            p->options.no_query = true;
+        } else if (!strcmp(token, "background-ping")) {
+            p->options.background_ping = true;
+        } else if (!strcmp(token, "no-digest")) {
+            p->options.no_digest = true;
+        } else if (!strcmp(token, "no-tproxy")) {
+            p->options.no_tproxy = true;
+        } else if (!strcmp(token, "multicast-responder")) {
+            p->options.mcast_responder = true;
+#if PEER_MULTICAST_SIBLINGS
+        } else if (!strcmp(token, "multicast-siblings")) {
+            p->options.mcast_siblings = true;
+#endif
+        } else if (!strncmp(token, "weight=", 7)) {
+            p->weight = xatoi(token + 7);
+        } else if (!strncmp(token, "basetime=", 9)) {
+            p->basetime = xatoi(token + 9);
+        } else if (!strcmp(token, "closest-only")) {
+            p->options.closest_only = true;
+        } else if (!strncmp(token, "ttl=", 4)) {
+            p->mcast.ttl = xatoi(token + 4);
+
+            if (p->mcast.ttl < 0)
+                p->mcast.ttl = 0;
+
+            if (p->mcast.ttl > 128)
+                p->mcast.ttl = 128;
+        } else if (!strcmp(token, "default")) {
+            p->options.default_parent = true;
+        } else if (!strcmp(token, "round-robin")) {
+            p->options.roundrobin = true;
+        } else if (!strcmp(token, "weighted-round-robin")) {
+            p->options.weighted_roundrobin = true;
+#if USE_HTCP
+        } else if (!strcmp(token, "htcp")) {
+            p->options.htcp = true;
+        } else if (!strncmp(token, "htcp=", 5) || !strncmp(token, "htcp-", 5)) {
+            /* Note: The htcp- form is deprecated, replaced by htcp= */
+            p->options.htcp = true;
+            char *tmp = xstrdup(token+5);
+            char *mode, *nextmode;
+            for (mode = nextmode = tmp; mode; mode = nextmode) {
+                nextmode = strchr(mode, ',');
+                if (nextmode) {
+                    *nextmode = '\0';
+                    ++nextmode;
+                }
+                if (!strcmp(mode, "no-clr")) {
+                    if (p->options.htcp_only_clr)
+                        fatalf("parse_peer: can't set htcp-no-clr and htcp-only-clr simultaneously");
+                    p->options.htcp_no_clr = true;
+                } else if (!strcmp(mode, "no-purge-clr")) {
+                    p->options.htcp_no_purge_clr = true;
+                } else if (!strcmp(mode, "only-clr")) {
+                    if (p->options.htcp_no_clr)
+                        fatalf("parse_peer: can't set htcp no-clr and only-clr simultaneously");
+                    p->options.htcp_only_clr = true;
+                } else if (!strcmp(mode, "forward-clr")) {
+                    p->options.htcp_forward_clr = true;
+                } else if (!strcmp(mode, "oldsquid")) {
+                    p->options.htcp_oldsquid = true;
+                } else {
+                    fatalf("invalid HTCP mode '%s'", mode);
+                }
+            }
+            safe_free(tmp);
+#endif
+        } else if (!strcmp(token, "no-netdb-exchange")) {
+            p->options.no_netdb_exchange = true;
+
+        } else if (!strcmp(token, "carp")) {
+            if (p->type != PEER_PARENT)
+                throw TextException(ToSBuf("non-parent carp cache_peer ", *p), Here());
+
+            p->options.carp = true;
+        } else if (!strncmp(token, "carp-key=", 9)) {
+            if (p->options.carp != true)
+                throw TextException(ToSBuf("carp-key specified on non-carp cache_peer ", *p), Here());
+            p->options.carp_key.set = true;
+            char *nextkey=token+strlen("carp-key="), *key=nextkey;
+            for (; key; key = nextkey) {
+                nextkey=strchr(key,',');
+                if (nextkey) ++nextkey; // skip the comma, any
+                if (0==strncmp(key,"scheme",6)) {
+                    p->options.carp_key.scheme = true;
+                } else if (0==strncmp(key,"host",4)) {
+                    p->options.carp_key.host = true;
+                } else if (0==strncmp(key,"port",4)) {
+                    p->options.carp_key.port = true;
+                } else if (0==strncmp(key,"path",4)) {
+                    p->options.carp_key.path = true;
+                } else if (0==strncmp(key,"params",6)) {
+                    p->options.carp_key.params = true;
+                } else {
+                    fatalf("invalid carp-key '%s'",key);
+                }
+            }
+        } else if (!strcmp(token, "userhash")) {
+#if USE_AUTH
+            if (p->type != PEER_PARENT)
+                throw TextException(ToSBuf("non-parent userhash cache_peer ", *p), Here());
+
+            p->options.userhash = true;
+#else
+            throw TextException(ToSBuf("missing authentication support; required for userhash cache_peer ", *p), Here());
+#endif
+        } else if (!strcmp(token, "sourcehash")) {
+            if (p->type != PEER_PARENT)
+                throw TextException(ToSBuf("non-parent sourcehash cache_peer ", *p), Here());
+
+            p->options.sourcehash = true;
+
+        } else if (!strcmp(token, "no-delay")) {
+#if USE_DELAY_POOLS
+            p->options.no_delay = true;
+#else
+            debugs(0, DBG_CRITICAL, "WARNING: cache_peer option 'no-delay' requires --enable-delay-pools");
+#endif
+        } else if (!strncmp(token, "login=", 6)) {
+            p->login = xstrdup(token + 6);
+            rfc1738_unescape(p->login);
+        } else if (!strcmp(token, "auth-no-keytab")) {
+            p->options.auth_no_keytab = 1;
+        } else if (!strncmp(token, "connect-timeout=", 16)) {
+            p->connect_timeout_raw = xatoi(token + 16);
+        } else if (!strncmp(token, "connect-fail-limit=", 19)) {
+            p->connect_fail_limit = xatoi(token + 19);
+#if USE_CACHE_DIGESTS
+        } else if (!strncmp(token, "digest-url=", 11)) {
+            p->digest_url = xstrdup(token + 11);
+#endif
+
+        } else if (!strcmp(token, "allow-miss")) {
+            p->options.allow_miss = true;
+        } else if (!strncmp(token, "max-conn=", 9)) {
+            p->max_conn = xatoi(token + 9);
+        } else if (!strncmp(token, "standby=", 8)) {
+            p->standby.limit = xatoi(token + 8);
+        } else if (!strcmp(token, "originserver")) {
+            p->options.originserver = true;
+        } else if (!strncmp(token, "name=", 5)) {
+            p->rename(token + 5);
+        } else if (!strncmp(token, "forceddomain=", 13)) {
+            safe_free(p->domain);
+            if (token[13])
+                p->domain = xstrdup(token + 13);
+
+        } else if (strncmp(token, "ssl", 3) == 0) {
+#if !USE_OPENSSL
+            debugs(0, DBG_CRITICAL, "WARNING: cache_peer option '" << token << "' requires --with-openssl");
+#else
+            p->secure.parse(token+3);
+#endif
+        } else if (strncmp(token, "tls-", 4) == 0) {
+            p->secure.parse(token+4);
+        } else if (strncmp(token, "tls", 3) == 0) {
+            p->secure.parse(token+3);
+        } else if (strcmp(token, "front-end-https") == 0) {
+            p->front_end_https = 1;
+        } else if (strcmp(token, "front-end-https=on") == 0) {
+            p->front_end_https = 1;
+        } else if (strcmp(token, "front-end-https=auto") == 0) {
+            p->front_end_https = 2;
+        } else if (strcmp(token, "connection-auth=off") == 0) {
+            p->connection_auth = 0;
+        } else if (strcmp(token, "connection-auth") == 0) {
+            p->connection_auth = 1;
+        } else if (strcmp(token, "connection-auth=on") == 0) {
+            p->connection_auth = 1;
+        } else if (strcmp(token, "connection-auth=auto") == 0) {
+            p->connection_auth = 2;
+        } else if (token[0] == '#') {
+            // start of a text comment. stop reading this line.
+            break;
+        } else {
+            debugs(3, DBG_PARSE_NOTE(DBG_IMPORTANT), "ERROR: Ignoring unknown cache_peer option '" << token << "'");
+        }
+    }
+
+    if (findCachePeerByName(p->name))
+        throw TextException(ToSBuf("cache_peer ", *p, " specified twice"), Here());
+
+    if (p->max_conn > 0 && p->max_conn < p->standby.limit)
+        throw TextException(ToSBuf("cache_peer ", *p, " max-conn=", p->max_conn,
+                                   " is lower than its standby=", p->standby.limit), Here());
+
+    if (p->weight < 1)
+        p->weight = 1;
+
+    if (p->connect_fail_limit < 1)
+        p->connect_fail_limit = 10;
+
+#if USE_CACHE_DIGESTS
+    if (!p->options.no_digest)
+        p->digest = new PeerDigest(p);
+#endif
+
+    if (p->secure.encryptTransport)
+        p->secure.parseOptions();
+
+    if (!*peers)
+        *peers = new CachePeers;
+
+    (*peers)->add(p);
+
+    p->index = (*peers)->size();
+
+    peerClearRRStart();
 }
 
 static void
@@ -2052,18 +2416,6 @@ free_peer(CachePeers **peers)
         p->clear();
         delete p;
         *peers = nullptr;
-    }
-}
-
-static void parse_peer(CachePeers **peers)
-{
-    try {
-        if (!*peers)
-            *peers = new CachePeers;
-        (*peers)->parse(LegacyParser);
-    } catch (...) {
-        free_peer(peers);
-        throw;
     }
 }
 
@@ -2177,7 +2529,7 @@ parse_hostdomaintype(void)
     char *domain = nullptr;
     while ((domain = ConfigParser::NextToken())) {
         auto *l = static_cast<NeighborTypeDomainList *>(xcalloc(1, sizeof(NeighborTypeDomainList)));
-        l->type = CachePeers::parseNeighborType(type);
+        l->type = parseNeighborType(type);
         l->domain = xstrdup(domain);
 
         NeighborTypeDomainList **L = nullptr;
@@ -2955,6 +3307,29 @@ dump_memcachemode(StoreEntry * entry, const char *name, SquidConfig &)
 }
 
 #include "cf_parser.cci"
+
+peer_t
+parseNeighborType(const char *s)
+{
+    if (!strcmp(s, "parent"))
+        return PEER_PARENT;
+
+    if (!strcmp(s, "neighbor"))
+        return PEER_SIBLING;
+
+    if (!strcmp(s, "neighbour"))
+        return PEER_SIBLING;
+
+    if (!strcmp(s, "sibling"))
+        return PEER_SIBLING;
+
+    if (!strcmp(s, "multicast"))
+        return PEER_MULTICAST;
+
+    debugs(15, DBG_CRITICAL, "WARNING: Unknown neighbor type: " << s);
+
+    return PEER_SIBLING;
+}
 
 #if USE_WCCPv2
 static void
